feat: integrate into srt and clean up old rop
