--- python/sglang/jit_kernel/rope_v2.py
@@ -1,92 +1,0 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-import torch
-
-from sglang.jit_kernel.utils import (
-    cache_once,
-    is_arch_support_pdl,
-    load_jit,
-    make_cpp_args,
-)
-
-if TYPE_CHECKING:
-    from tvm_ffi.module import Module
-
-
-@cache_once
-def _jit_fused_rope_module(is_neox: bool, rope_dim: int, dtype: torch.dtype) -> Module:
-    args = make_cpp_args(is_neox, rope_dim, is_arch_support_pdl(), dtype)
-    return load_jit(
-        "fused_rope_v2",
-        *args,
-        cuda_files=["elementwise/rope_v2.cuh"],
-        cuda_wrappers=[
-            ("run_rope", f"FusedRopeKernel<{args}>::run"),
-            ("run_rope_store", f"FusedRopeKernel<{args}>::run_fused"),
-        ],
-    )
-
-
-def fused_rope_inplace(
-    q: torch.Tensor,
-    k: torch.Tensor,
-    cos_sin_cache: torch.Tensor,
-    positions: torch.Tensor,
-    *,
-    is_neox: bool = True,
-    rope_dim: int = 0,
-) -> None:
-    """
-    Fused inplace rotary position embedding for query and key tensors.
-
-    Args:
-        q: Query tensor of shape [num_tokens, num_qo_heads, rope_dim].
-        k: Key tensor of shape [num_tokens, num_kv_heads, rope_dim].
-        cos_sin_cache: Cosine/sine cache of shape [max_position, rope_dim],
-            where the first half along dim=-1 is cos and the second half is sin.
-            Must be float32.
-        positions: Position indices of shape [num_tokens], int32 or int64.
-        is_neox: Whether to use GPT-NeoX style (True) or GPT-J interleaved style (False).
-        rope_dim: Rotary embedding dimension. Defaults to cos_sin_cache.size(-1).
-    """
-    rope_dim = rope_dim or cos_sin_cache.size(-1)
-    module = _jit_fused_rope_module(is_neox, rope_dim, q.dtype)
-    module.run_rope(q, k, cos_sin_cache, positions)
-
-
-def fused_rope_inplace_with_kvcache(
-    q: torch.Tensor,
-    k: torch.Tensor,
-    v: torch.Tensor,
-    k_cache: torch.Tensor,
-    v_cache: torch.Tensor,
-    cos_sin_cache: torch.Tensor,
-    positions: torch.Tensor,
-    out_loc: torch.Tensor,
-    *,
-    is_neox: bool = True,
-    rope_dim: int = 0,
-) -> None:
-    """
-    Fused inplace RoPE + KV cache store.
-
-    Applies rotary position embedding to q inplace. For k, applies RoPE and
-    stores the result in k_cache. The original v is also stored in v_cache.
-
-    Args:
-        q: Query tensor of shape [num_tokens, num_qo_heads, head_dim].
-        k: Key tensor of shape [num_tokens, num_kv_heads, head_dim].
-        v: Value tensor of shape [num_tokens, num_kv_heads, head_dim].
-        k_cache: Key cache of shape [cache_size, num_kv_heads * head_dim].
-        v_cache: Value cache of shape [cache_size, num_kv_heads * head_dim].
-        cos_sin_cache: Cosine/sine cache of shape [max_position, rope_dim], float32.
-        positions: Position indices of shape [num_tokens], int32 or int64.
-        out_loc: Cache write locations of shape [num_tokens], same dtype as positions.
-        is_neox: Whether to use GPT-NeoX style (True) or GPT-J interleaved (False).
-        rope_dim: Rotary embedding dimension. Defaults to cos_sin_cache.size(-1).
-    """
-    rope_dim = rope_dim or cos_sin_cache.size(-1)
-    module = _jit_fused_rope_module(is_neox, rope_dim, q.dtype)
-    module.run_rope_store(q, k, v, k_cache, v_cache, cos_sin_cache, positions, out_loc)
