+++ python/sglang/srt/models/utils.py
@@ -119,20 +119,18 @@ def create_fused_set_kv_buffer_arg(
     layer: RadixAttention,
     forward_batch: ForwardBatch,
 ):
-    from sgl_kernel import FusedSetKVBufferArg
+    from sglang.jit_kernel.rope import FusedSetKVBufferArg
 
     layer_id = layer.layer_id
     token_to_kv_pool = forward_batch.token_to_kv_pool
 
     k_buffer = token_to_kv_pool.get_key_buffer(layer_id)
     v_buffer = token_to_kv_pool.get_value_buffer(layer_id)
-
+    assert layer.k_scale is None and layer.v_scale is None, "scale not supported"
     return FusedSetKVBufferArg(
         value=value,
         k_buffer=k_buffer.view(k_buffer.shape[0], -1),
         v_buffer=v_buffer.view(v_buffer.shape[0], -1),
-        k_scale=layer.k_scale,
-        v_scale=layer.v_scale,
         cache_loc=forward_batch.out_cache_loc,
     )
